Rename üackage name for catalouge plugin
