+++ Kitodo/src/main/java/org/goobi/production/plugin/UnspecificPlugin.java
@@ -18,7 +18,7 @@ import java.util.Locale;
 import java.util.Map;
 
 import org.goobi.production.enums.PluginType;
-import org.goobi.production.plugin.CataloguePlugin.CataloguePlugin;
+import org.goobi.production.plugin.catalogue.CataloguePlugin;
 import org.kitodo.exceptions.NotImplementedException;
 import org.kitodo.exceptions.PluginException;
 import org.kitodo.exceptions.UnreachableCodeException;
