+++ Kitodo/src/main/java/org/kitodo/forms/ProzesskopieForm.java
@@ -40,10 +40,10 @@ import org.apache.commons.lang.SystemUtils;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.goobi.production.cli.helper.WikiFieldHelper;
-import org.goobi.production.plugin.CataloguePlugin.CataloguePlugin;
-import org.goobi.production.plugin.CataloguePlugin.Hit;
-import org.goobi.production.plugin.CataloguePlugin.QueryBuilder;
 import org.goobi.production.plugin.PluginLoader;
+import org.goobi.production.plugin.catalogue.CataloguePlugin;
+import org.goobi.production.plugin.catalogue.Hit;
+import org.goobi.production.plugin.catalogue.QueryBuilder;
 import org.jdom.JDOMException;
 import org.kitodo.api.ugh.DigitalDocumentInterface;
 import org.kitodo.api.ugh.DocStructInterface;
